Move extensionPoints into a package
This is to make it clearer where to start for someone new to this API
