+++ source/extensionPoints/util.py
@@ -1,25 +1,21 @@
-#extensionPoints.py
+#util.py
 #A part of NonVisual Desktop Access (NVDA)
 #Copyright (C) 2017 NV Access Limited
 #This file is covered by the GNU General Public License.
 #See the file COPYING for more details.
 
-"""Framework to enable extensibility at specific points in the code.
-This allows interested parties to register to be notified when some action occurs
-or to modify a specific kind of data.
-For example, you might wish to notify about a configuration profile switch
-or allow modification of spoken messages before they are passed to the synthesizer.
-See the L{Action} and L{Filter} classes.
+"""Utilities used withing the extension points framework. Generally it is expected that the class in __init__.py are
+used, however for more advanced requirements these utilities can be used directly.
 """
-
 import weakref
 import collections
 import inspect
-from logHandler import log
+
 
 class AnnotatableWeakref(weakref.ref):
 	"""A weakref.ref which allows annotation with custom attributes.
 	"""
+
 
 class BoundMethodWeakref(object):
 	"""Weakly references a bound instance method.
@@ -60,6 +56,7 @@ def _getHandlerKey(handler):
 	if inst:
 		return (id(inst), id(handler.__func__))
 	return id(handler)
+
 
 class HandlerRegistrar(object):
 	"""Base class to Facilitate registration and unregistration of handler functions.
@@ -117,6 +114,7 @@ class HandlerRegistrar(object):
 			if not handler:
 				continue # Died.
 			yield handler
+
 
 def callWithSupportedKwargs(func, *args, **kwargs):
 	"""Call a function with only the keyword arguments it supports.
@@ -184,125 +182,3 @@ def callWithSupportedKwargs(func, *args, **kwargs):
 		if kwarg not in supportedKwargs:
 			del kwargs[kwarg]
 	return func(*args, **kwargs)
-
-class Action(HandlerRegistrar):
-	"""Allows interested parties to register to be notified when some action occurs.
-	For example, this might be used to notify that the configuration profile has been switched.
-
-	First, an Action is created:
-
-	>>> somethingHappened = extensionPoints.Action()
-
-	Interested parties then register to be notified about this action, see
-	L{HandleRegistrar.register} docstring for details of the type of handlers that can be
-	registered:
-
-	>>> def onSomethingHappened(someArg=None):
-		... 	print(someArg)
-	...
-	>>> somethingHappened.register(onSomethingHappened)
-
-	When the action is performed, register handlers are notified, see C@{callWithSupportedKwarg}
-	for how args passed to notify are mapped to the handler:
-
-	>>> somethingHappened.notify(someArg=42)
-	"""
-
-	def notify(self, **kwargs):
-		"""Notify all registered handlers that the action has occurred.
-		@param kwargs: Arguments to pass to the handlers.
-		"""
-		for handler in self.handlers:
-			try:
-				callWithSupportedKwargs(handler, **kwargs)
-			except:
-				log.exception("Error running handler %r for %r" % (handler, self))
-
-class Filter(HandlerRegistrar):
-	"""Allows interested parties to register to modify a specific kind of data.
-	For example, this might be used to allow modification of spoken messages before they are passed to the synthesizer.
-
-	First, a Filter is created:
-
-	>>> messageFilter = extensionPoints.Filter()
-
-	Interested parties then register to filter the data, see
-	L{HandleRegistrar.register} docstring for details of the type of handlers that can be
-	registered:
-
-	>>> def filterMessage(message, someArg=None):
-	... 	return message + " which has been filtered."
-	...
-	>>> messageFilter.register(filterMessage)
-
-	When filtering is desired, all registered handlers are called to filter the data:
-
-	>>> messageFilter.apply("This is a message", someArg=42)
-	'This is a message which has been filtered'
-	"""
-
-	def apply(self, value, **kwargs):
-		"""Pass a value to be filtered through all registered handlers.
-		The value is passed to the first handler
-		and the return value from that handler is passed to the next handler.
-		This process continues for all handlers until the final handler.
-		The return value from the final handler is returned to the caller.
-		@param value: The value to be filtered.
-		@param kwargs: Arguments to pass to the handlers.
-		@return: The filtered value.
-		"""
-		for handler in self.handlers:
-			try:
-				value = callWithSupportedKwargs(handler, value, **kwargs)
-			except:
-				log.exception("Error running handler %r for %r" % (handler, self))
-		return value
-
-class Decider(HandlerRegistrar):
-	"""Allows interested parties to participate in deciding whether something
-	should be done.
-	For example, input gestures are normally executed,
-	but this might be used to prevent their execution
-	under specific circumstances such as when controlling a remote system.
-
-	First, a Decider is created:
-
-	>>> doSomething = extensionPoints.Decider()
-
-	Interested parties then register to participate in the decision, see
-	L{HandleRegistrar.register} docstring for details of the type of handlers that can be
-	registered:
-
-	>>> def shouldDoSomething(someArg=None):
-	... 	return False
-	...
-	>>> doSomething.register(shouldDoSomething)
-
-	When the decision is to be made, registered handlers are called until
-	a handler returns False:
-
-	>>> doSomething.decide(someArg=42)
-	False
-
-	If there are no handlers or all handlers return True,
-	the return value is True.
-	"""
-
-	def decide(self, **kwargs):
-		"""Call handlers to make a decision.
-		If a handler returns False, processing stops
-		and False is returned.
-		If there are no handlers or all handlers return True, True is returned.
-		@param kwargs: Arguments to pass to the handlers.
-		@return: The decision.
-		@rtype: bool
-		"""
-		for handler in self.handlers:
-			try:
-				decision = callWithSupportedKwargs(handler, **kwargs)
-			except:
-				log.exception("Error running handler %r for %r" % (handler, self))
-				continue
-			if not decision:
-				return False
-		return True
